[Reorg] NIF stream classes to own file
