+++ src/data/nifvalue.h
@@ -41,8 +41,6 @@ THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #include <QString>
 #include <QVariant>
 
-#include <memory>
-
 
 //! @file nifvalue.h NifValue, NifIStream, NifOStream, NifSStream
 
@@ -398,104 +396,6 @@ protected:
 
 Q_DECLARE_METATYPE( NifValue )
 
-
-class BaseModel;
-class NifItem;
-
-class QDataStream;
-class QIODevice;
-
-//! An input stream that reads a file into a model.
-class NifIStream final
-{
-	Q_DECLARE_TR_FUNCTIONS( NifIStream )
-
-public:
-	NifIStream( BaseModel * m, QIODevice * d ) : model( m ), device( d )
-	{
-		init();
-	}
-
-	//! Reads a NifValue from the underlying device. Returns true if successful.
-	bool read( NifValue & );
-
-private:
-	//! The model that data is being read into.
-	BaseModel * model;
-	//! The underlying device that data is being read from.
-	QIODevice * device;
-	//! The data stream that is wrapped around the device (simplifies endian conversion)
-	std::unique_ptr<QDataStream> dataStream;
-
-	//! Initialises the stream.
-	void init();
-
-	//! Whether a boolean is 32-bit.
-	bool bool32bit = false;
-	//! Whether link adjustment is required.
-	bool linkAdjust = false;
-	//! Whether string adjustment is required.
-	bool stringAdjust = false;
-	//! Whether the model is big-endian
-	bool bigEndian = false;
-
-	//! The maximum length of a string that can be read.
-	int maxLength = 0x8000;
-};
-
-
-//! An output stream that writes a model to a file.
-class NifOStream final
-{
-	Q_DECLARE_TR_FUNCTIONS( NifOStream )
-
-public:
-	NifOStream( const BaseModel * n, QIODevice * d ) : model( n ), device( d ) { init(); }
-
-	//! Writes a NifValue to the underlying device. Returns true if successful.
-	bool write( const NifValue & );
-
-private:
-	//! The model that data is being read from.
-	const BaseModel * model;
-	//! The underlying device that data is being written to.
-	QIODevice * device;
-
-	//! Initialises the stream.
-	void init();
-
-	//! Whether a boolean is 32-bit.
-	bool bool32bit = false;
-	//! Whether link adjustment is required.
-	bool linkAdjust = false;
-	//! Whether string adjustment is required.
-	bool stringAdjust = false;
-	//! Whether the model is big-endian
-	bool bigEndian = false;
-};
-
-
-//! A stream that determines the size of values in a model.
-class NifSStream final
-{
-public:
-	NifSStream( const BaseModel * n ) : model( n ) { init(); }
-
-	//! Determine the size of a given NifValue.
-	int size( const NifValue & );
-
-private:
-	//! The model that values are being sized for.
-	const BaseModel * model;
-
-	//! Initialises the stream.
-	void init();
-
-	//! Whether booleans are 32-bit or not.
-	bool bool32bit;
-	//! Whether string adjustment is required.
-	bool stringAdjust;
-};
 
 
 // Inlines
