+++ src/model/nifmodel.cpp
@@ -31,6 +31,7 @@ THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 ***** END LICENCE BLOCK *****/
 
 #include "nifmodel.h"
+#include "nifstream.h"
 
 #include "niftypes.h"
 #include "spellbook.h"
